requirements.txt
python-dotenv
pdfplumber
playwright
beautifulsoup4
markdown
langchain-core
langchain-google-genai
langchain-openai
langgraph
pydantic
fastapi
uvicorn[standard]
